WIP DefaultRouteMonitor implementation
This is a work-in-progress commit which shows one potential
implementation of the DefaultRouteMonitor
However there are several problems with this implementation that should
be thought about. There is a significant amount of potentially
unnecessary unsafe code for example.
The formatting, naming and other refinement related things in this
commit will be changed.